+++ compose.yml
@@ -28,7 +28,6 @@ x-service-template: &service-template
     - node_modules_ghost_post-revisions:/home/ghost/ghost/post-revisions/node_modules:delegated
     - node_modules_ghost_prometheus-metrics:/home/ghost/ghost/prometheus-metrics/node_modules:delegated
     - node_modules_ghost_security:/home/ghost/ghost/security/node_modules:delegated
-    - node_modules_ghost_tiers:/home/ghost/ghost/tiers/node_modules:delegated
     - node_modules_ghost_webmentions:/home/ghost/ghost/webmentions/node_modules:delegated
     - node_modules_apps_admin-x-activitypub:/home/ghost/apps/admin-x-activitypub/node_modules:delegated
     - node_modules_apps_admin-x-design-system:/home/ghost/apps/admin-x-design-system/node_modules:delegated
@@ -197,7 +196,6 @@ volumes:
   node_modules_ghost_post-revisions: {}
   node_modules_ghost_prometheus-metrics: {}
   node_modules_ghost_security: {}
-  node_modules_ghost_tiers: {}
   node_modules_ghost_webmentions: {}
   node_modules_apps_admin-x-activitypub: {}
   node_modules_apps_admin-x-design-system: {}
