+++ ghost/core/core/server/services/tiers/TierNameChangeEvent.js
@@ -1,6 +1,6 @@
 /**
  * @typedef {object} TierNameChangeEventData
- * @prop {Tier} tier
+ * @prop {import('./Tier')} tier
  */
 
 class TierNameChangeEvent {
