+++ ghost/core/package.json
@@ -118,7 +118,6 @@
     "@tryghost/security": "0.0.0",
     "@tryghost/social-urls": "0.1.47",
     "@tryghost/string": "0.2.13",
-    "@tryghost/tiers": "0.0.0",
     "@tryghost/tpl": "0.1.34",
     "@tryghost/url-utils": "4.4.9",
     "@tryghost/validator": "0.2.16",
