+++ ghost/core/test/unit/server/services/members/members-api/services/payments.test.js
@@ -1,7 +1,8 @@
 const assert = require('assert/strict');
 const sinon = require('sinon');
-const knex = require('knex');
-const {Tier} = require('@tryghost/tiers');
+const knex = require('knex').default;
+
+const Tier = require('../../../../../../../core/server/services/tiers/Tier');
 
 const PaymentsService = require('../../../../../../../core/server/services/members/members-api/services/PaymentsService');
