--- ghost/tiers/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "@tryghost/tiers",
-  "version": "0.0.0",
-  "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/tiers",
-  "author": "Ghost Foundation",
-  "private": true,
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.8.2"
-  },
-  "dependencies": {
-    "@tryghost/errors": "1.3.7",
-    "@tryghost/string": "0.2.13",
-    "@tryghost/tpl": "0.1.34",
-    "bson-objectid": "2.0.4"
-  }
-}
